Added routing for full and slim UI versions
