+++ package.json
@@ -15,6 +15,7 @@
     "react-calendar-timeline": "^0.14.0",
     "react-dom": "^15.5.4",
     "react-flexbox-grid": "^1.1.3",
+    "react-router-dom": "^4.1.1",
     "react-tap-event-plugin": "^2.0.1",
     "slate": "^0.20.1",
     "slate-auto-replace": "^0.5.0"
