+++ src/views/FullApp.css
@@ -1,0 +1,63 @@
+.FullApp {
+    height: 100vh;
+}
+
+/* Restricts app to max fixed width for very wide screens. */
+.FullApp-content {
+    max-width: 1600px;
+}
+
+.container-fluid {
+    padding-left: 12px;
+    padding-right: 12px;
+}
+
+.dashboard-panel {
+    margin-top: 16px;
+    text-align: left;
+}
+
+.trio {
+    min-height: 810px !important;
+}
+
+.panel-content {
+    padding: 30px;
+}
+
+.divider {
+    margin: 20px 0px !important;
+    clear: both;
+}
+
+h1 {
+    margin: 0px;
+    font-size: 1.25rem;
+}
+
+h2 {
+    margin: 0px;
+    font-size: 0.9rem;
+    font-weight: lighter;
+}
+
+h3 {
+    margin: 0px;
+    font-size: 0.8rem;
+    font-weight: lighter;
+}
+
+
+.FullApp-content ::-webkit-scrollbar {
+    width: 12px;
+}
+
+.FullApp-content ::-webkit-scrollbar-track {
+    -webkit-box-shadow: inset 0 0 1px rgba(0,0,0,0);
+    border-radius: 10px;
+}
+
+.FullApp-content ::-webkit-scrollbar-thumb {
+    border-radius: 10px;
+    -webkit-box-shadow: inset 0 0 5px rgba(0,0,0,0.5);
+}
